% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_extract_readability.R
\name{cas_extract_readability}
\alias{cas_extract_readability}
\title{Extract text and metadata from an html string with \code{readability.js}}
\usage{
cas_extract_readability(html, url, candidates = 5L, threshold = 500L)
}
\arguments{
\item{threshold}{}
}
\description{
Extract text and metadata from an html string with \code{readability.js}
}
