% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_extract_is_readable.R
\name{cas_extract_readability_check}
\alias{cas_extract_readability_check}
\title{Check if a an html input can be successfully parsed through \code{readability.js}}
\usage{
cas_extract_readability_check(html, min_content_length = 140, min_score = 20)
}
\arguments{
\item{min_score}{}
}
\description{
Check if a an html input can be successfully parsed through \code{readability.js}
}
